+++ CHANGELOG.md
@@ -21,6 +21,12 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### Fixes
+
+* Option Types
+  * DataPath
+    - added missing `MakeDataPathOptions`
+
 ## v0.1.8
 
 Released Monday, 22nd June 2020.